:recycle: Helper | Move all table helpers to seperate files
