+++ apps/web/src/components/clientPages/validators.tsx
@@ -4,7 +4,6 @@ import { Button, DonutChart, FlexGrid } from '@repo/ui/atoms';
 import { NextValidators } from '@repo/ui/molecules';
 import { useEffect, useState } from 'react';
 import { validatorsTableHead } from '../../utils/helpers/tableHeaders.tsx';
-import { createValidatorsRows } from '../../utils/helpers/helper.tsx';
 import { useSocketStore } from '../../store/socketStore.ts';
 import { ChartDataType, ValidatorType } from '../../utils/types.ts';
 import {
@@ -15,8 +14,9 @@ import {
 import { useGatewayClientStore } from '../../store/clientStore.ts';
 import { NextValidatorType, TableCellType } from '@repo/ui/types';
 import { useBasePath } from '../../utils/hooks/useBasePath.ts';
-import {useChainNetworkStore} from "../../store/chainNetworkStore.ts";
-import {useIsHorizontallyScrolled} from "../../utils/hooks/useIsHorizontallyScrolled.ts";
+import { useChainNetworkStore } from '../../store/chainNetworkStore.ts';
+import { useIsHorizontallyScrolled } from '../../utils/hooks/useIsHorizontallyScrolled.ts';
+import { createValidatorsRows } from '../../utils/helpers/TableHelpers/accountTableHelper.tsx';
 
 export const Validators = () => {
   const [validators, setValidators] = useState<ValidatorType[]>([]);
