--- apps/web/src/utils/helpers/helper.tsx
@@ -1,1191 +1,0 @@
-import {
-  BlockType,
-  EventsType,
-  FavouriteType,
-  StakesCalculatorPeriodType,
-  AccountType,
-  StakeType,
-  TransactionType,
-  ValidatorType,
-  NodeType,
-  TokenType,
-  NftType,
-  AppsType,
-  CombinedAppsType,
-} from '../types.ts';
-import { cls, fromNowFormatter, replaceColonWithSpace, shortString } from '@repo/ui/utils';
-import { ImageName, TxDataPopover } from '@repo/ui/molecules';
-import {
-  Badge,
-  JsonViewer,
-  Typography,
-  NotificationIcon,
-  StatusBadge,
-  StatusIcon,
-  KeyValueComponent,
-  IconButton,
-  TokenCard,
-  FlexGrid,
-  Button,
-} from '@repo/ui/atoms';
-import { Link } from '@repo/ui/atoms';
-import React from 'react';
-import { commandColors, decimals } from '../constants.tsx';
-import { convertKLYToBeddows, getSeedRevealFromAssets, getTableSkeletons } from './dataHelpers.tsx';
-import {
-  eventsTableHead,
-  validatorsTableHead,
-  validatorBlocksTableHead,
-  validatorStakeIncomingTableHead,
-  validatorStakeOutgoingTableHead,
-  transactionTableHead,
-  blockTableHead,
-  stakesOverviewTableHead,
-  favouritesTableHead,
-  chainsTableHead,
-} from './tableHeaders.tsx';
-import { ChainType, DataType } from '@repo/ui/types';
-import { formatCommission, getAmountFromTx } from './dataHelpers.tsx';
-import { Currency } from '../../components/currency.tsx';
-import { FormattedValue } from '../../components/formattedValue.tsx';
-import { AppType } from 'next/app';
-
-export const createTransactionRows = (
-  transactions: TransactionType[],
-  loading: boolean,
-  currentChain: ChainType,
-  chains: ChainType[],
-  copyTooltipText: string,
-  setCopyTooltipText: (text: string) => void,
-  basePath: string,
-  statusOfTransaction?: boolean,
-) => {
-  const columnCount = transactionTableHead(() => '', '', '').length;
-
-  const chainLogo = currentChain?.logo;
-  const getChainLogo = (chainID: string) => {
-    const fromChain = chains.find((chain) => {
-      const isTheChain = chain.chainID === chainID;
-      return isTheChain;
-    });
-    const logo = fromChain?.logo.png;
-    return logo;
-  };
-
-  return !loading
-    ? transactions?.length > 0
-      ? transactions?.map((transaction) => {
-          const cells = [
-            {
-              children: (
-                <KeyValueComponent
-                  contentValue={
-                    <FormattedValue
-                      copy
-                      format={'address'}
-                      link={`/transactions/${transaction.id}`}
-                      showCopyOnHover
-                      value={transaction?.id}
-                    />
-                  }
-                  keyValue={<StatusIcon status={transaction.executionStatus} />}
-                />
-              ),
-            },
-            {
-              children: (
-                <FormattedValue
-                  copy
-                  format={'number'}
-                  showCopyOnHover
-                  typographyProps={{ color: 'onBackgroundLow' }}
-                  value={transaction.block.height}
-                />
-              ),
-              className: 'group/child',
-            },
-            {
-              children: (
-                <FormattedValue
-                  format={'fromNow'}
-                  typographyProps={{ color: 'onBackgroundLow' }}
-                  value={transaction.block.timestamp}
-                />
-              ),
-            },
-            {
-              children: (
-                <Badge
-                  colorVariant={commandColors[transaction.command]}
-                  label={replaceColonWithSpace(`${transaction?.module}:${transaction?.command}`)}
-                />
-              ),
-            },
-            {
-              children: (
-                <div className="">
-                  <FormattedValue
-                    format={'account'}
-                    value={transaction.sender}
-                    accountIconComponent={
-                      transaction.receivingChainID ? (
-                        <img
-                          alt="Chain Icon"
-                          className="absolute -left-2 bottom-4 rounded-full"
-                          height={20}
-                          src={chainLogo.png}
-                          width={20}
-                        />
-                      ) : null
-                    }
-                  />
-                </div>
-              ),
-            },
-            {
-              children: transaction?.recipient ? (
-                <div className="">
-                  <FormattedValue
-                    format={'account'}
-                    value={transaction.recipient}
-                    accountIconComponent={
-                      transaction.receivingChainID ? (
-                        <img
-                          alt="Chain Icon"
-                          className="absolute -left-2 bottom-4 rounded-full"
-                          height={20}
-                          src={getChainLogo(transaction.receivingChainID)}
-                          width={20}
-                        />
-                      ) : null
-                    }
-                  />
-                </div>
-              ) : (
-                '-'
-              ),
-            },
-            {
-              children: (
-                <Currency
-                  amount={getAmountFromTx(transaction)}
-                  className={'align-middle'}
-                  color={'onBackgroundLow'}
-                  decimals={decimals}
-                  variant={'paragraph-sm'}
-                />
-              ),
-              className: 'text-right',
-            },
-            {
-              children: (
-                <Currency
-                  amount={transaction?.fee}
-                  className={'align-middle'}
-                  color={'onBackgroundLow'}
-                  decimals={5}
-                  variant={'paragraph-sm'}
-                />
-              ),
-              className: 'text-right',
-            },
-          ];
-
-          if (statusOfTransaction) {
-            cells.splice(4, 0, {
-              children: <StatusBadge status={transaction.executionStatus} />,
-            });
-          }
-
-          return {
-            rowDetails: (
-              <TxDataPopover
-                txData={{
-                  status: transaction?.executionStatus || 'pending',
-                  data: transaction?.params?.data,
-                  nonce: transaction?.nonce,
-                }}
-              />
-            ),
-            cells,
-          };
-        })
-      : [
-          {
-            cells: [
-              {
-                children: <Typography>{'No transactions found'}</Typography>,
-                colSpan: columnCount,
-              },
-            ],
-          },
-        ]
-    : getTableSkeletons(columnCount);
-};
-
-export const createEventsRows = (events: EventsType[], loading: boolean) => {
-  const columnCount = eventsTableHead.length;
-
-  return !loading
-    ? events?.map((event) => {
-        return {
-          rowDetails: (
-            <JsonViewer
-              className={'!border-0'}
-              copy
-              data={{ data: event.data as unknown as DataType }}
-              startOpen
-            />
-          ),
-          cells: [
-            {
-              children: (
-                <FormattedValue
-                  format={'string'}
-                  typographyProps={{ color: 'onBackgroundHigh' }}
-                  value={event.module}
-                />
-              ),
-              className: 'desktop:w-1/5',
-            },
-            {
-              children: (
-                <FormattedValue
-                  format={'string'}
-                  typographyProps={{ color: 'onBackgroundHigh' }}
-                  value={event.name}
-                />
-              ),
-            },
-          ],
-        };
-      })
-    : getTableSkeletons(columnCount);
-};
-
-export const createValidatorsRows = (
-  validators: ValidatorType[],
-  loading: boolean,
-  isScrolled: boolean,
-  stakingRewards = false,
-  stakingCalculatorProps:
-    | {
-        stakingCalculatorAmount: number;
-        stakingCalculatorPeriod: StakesCalculatorPeriodType;
-        totalActiveStake: bigint;
-      }
-    | undefined = {
-    stakingCalculatorAmount: 1000,
-    stakingCalculatorPeriod: 'day',
-    totalActiveStake: BigInt(0),
-  },
-) => {
-  const columnCount = validatorsTableHead(() => '', '', '').length;
-  const { stakingCalculatorAmount, stakingCalculatorPeriod, totalActiveStake } =
-    stakingCalculatorProps;
-
-  const calculateReward = (validator: ValidatorType) => {
-    if (!totalActiveStake) {
-      return {
-        resultPerBlock: '0',
-        inputStake: 0,
-        capacity: 0,
-        newBlockReward: '0',
-      };
-    }
-
-    const newStake = BigInt(stakingCalculatorAmount) * BigInt(1_0000_0000);
-    const newStakeFloat = parseFloat(newStake.toString(10));
-    const newTotalStake = totalActiveStake + BigInt(newStake);
-    const newTotalStakeFloat = parseFloat(newTotalStake.toString(10));
-    const newValidatorWeight =
-      parseFloat(validator.validatorWeight) + newStakeFloat > parseFloat(validator.selfStake) * 10
-        ? parseFloat(validator.selfStake) * 10
-        : parseFloat(validator.validatorWeight) + newStakeFloat;
-    const Commission = validator.commission / 100;
-    const newVoteShare =
-      parseFloat(newStake.toString(10)) / (parseFloat(validator.totalStake) + newStakeFloat);
-    const share = (100 - Commission) / 100;
-    const rewardPerBlock = (newValidatorWeight / newTotalStakeFloat) * 90.9 + 0.1;
-    const newBlockReward = newVoteShare * rewardPerBlock * share;
-    const RewardPerBlock = parseInt(validator.blockReward, 10);
-    const stakingCalculatorBeddows = stakingCalculatorAmount * 100000000;
-    const Stake = parseFloat(validator.totalStake) + stakingCalculatorBeddows;
-    const capacity =
-      (parseFloat(validator.totalStake) / (parseFloat(validator.selfStake) * 10)) * 100;
-
-    const stakersRewardPerBlock = (RewardPerBlock: any, Commission: any, Stake: any) =>
-      RewardPerBlock * (1 - Commission / 100) * (stakingCalculatorBeddows / Stake);
-    const resultPerBlock =
-      validator.status === 'active'
-        ? parseInt(stakersRewardPerBlock(RewardPerBlock, Commission, Stake).toString()).toString()
-        : '0';
-    return {
-      resultPerBlock,
-      inputStake: stakingCalculatorBeddows,
-      capacity,
-      newBlockReward:
-        validator.status === 'active'
-          ? newBlockReward > 90.9 * 0.1 + 0.1
-            ? convertKLYToBeddows((90.9 * 0.1 + 0.1).toFixed(8))
-            : convertKLYToBeddows(newBlockReward.toFixed(8))
-          : '0',
-    };
-  };
-
-  return !loading
-    ? validators?.map((validator) => {
-        const { inputStake, newBlockReward } = calculateReward(validator);
-
-        const getWeightPercents = (validator: ValidatorType) => {
-          const rawPercents = Number(
-            ((Number(validator?.totalStake || 0) / Number(validator?.selfStake || 1)) * 10).toFixed(
-              2,
-            ),
-          );
-          if (isNaN(rawPercents)) {
-            return 0;
-          } else if (!isFinite(rawPercents)) {
-            return 0;
-          } else {
-            return rawPercents;
-          }
-        };
-        const weightPercents = getWeightPercents(validator);
-
-        const resultPerPeriod =
-          stakingCalculatorPeriod === 'block'
-            ? newBlockReward
-            : stakingCalculatorPeriod === 'day'
-              ? (parseInt(newBlockReward) * 84).toString(10)
-              : stakingCalculatorPeriod === 'month'
-                ? (parseInt(newBlockReward) * 2516).toString(10)
-                : stakingCalculatorPeriod === 'year'
-                  ? (parseInt(newBlockReward) * 2516 * 12).toString(10)
-                  : (parseInt(newBlockReward) * 2516 * 12).toString(10);
-
-        const APR = ((parseInt(newBlockReward) * 2516 * 12) / inputStake) * 100;
-
-        return {
-          cells: [
-            {
-              children: (
-                <FormattedValue
-                  accountIconComponent={
-                    <NotificationIcon
-                      className="absolute -translate-x-3 -translate-y-3"
-                      notificationValue={validator?.rank < 999 ? validator?.rank : ''}
-                      size="lg"
-                    />
-                  }
-                  format={'account'}
-                  value={validator?.account}
-                />
-              ),
-              className: cls([!stakingRewards && 'sticky left-0 bg-background group-hover:bg-backgroundSecondary', !stakingRewards && isScrolled ? 'shadow-border-r' : '']),
-            },
-            {
-              children: (
-                <StatusBadge
-                  nextAllocatedTime={
-                    validator.nextAllocatedTime
-                      ? fromNowFormatter(validator.nextAllocatedTime * 1000, 'DD MMM YYYY')
-                      : undefined
-                  }
-                  status={validator.status}
-                />
-              ),
-            },
-            stakingRewards
-              ? null
-              : {
-                  children: <FormattedValue format={'number'} value={validator?.generatedBlocks} />,
-                  className: 'text-right',
-                },
-
-            {
-              children: (
-                <div className="flex flex-col items-end">
-                  <Currency
-                    amount={validator?.validatorWeight}
-                    className="font-semibold"
-                    decimals={0}
-                  />
-                  <FormattedValue
-                    format={'percentage'}
-                    tooltip={{
-                      text: 'Stake capacity',
-                      placement: 'top',
-                    }}
-                    typographyProps={{
-                      color: weightPercents > 100 ? 'error' : 'onBackgroundLow',
-                    }}
-                    value={weightPercents}
-                  />
-                </div>
-              ),
-            },
-            stakingRewards
-              ? null
-              : {
-                  children: (
-                    <div className="flex flex-col items-end">
-                      <Currency
-                        amount={validator?.totalStake}
-                        className="font-semibold"
-                        decimals={0}
-                      />
-                      <FormattedValue
-                        format={'currency'}
-                        tooltip={{
-                          text: 'Self stake',
-                          placement: 'top',
-                        }}
-                        currencyProps={{
-                          decimals: 0,
-                          className: 'text-caption text-onBackgroundLow',
-                        }}
-                        value={validator?.selfStake}
-                      />
-                    </div>
-                  ),
-                  className: 'text-right',
-                },
-
-            {
-              children: (
-                <FormattedValue
-                  format={'percentage'}
-                  currencyProps={{ variant: 'paragraph-sm' }}
-                  value={formatCommission(validator?.commission)}
-                />
-              ),
-              className: 'text-right',
-            },
-            {
-              children: (
-                <Currency amount={validator.totalRewards} color={'onBackgroundLow'} decimals={0} />
-              ),
-              className: 'text-right',
-            },
-            {
-              children: (
-                <Currency amount={validator.blockReward} color={'onBackgroundLow'} decimals={5} />
-              ),
-              className: 'text-right',
-            },
-          ].filter(Boolean),
-        };
-      })
-    : getTableSkeletons(columnCount);
-};
-
-export const createValidatorIncomingStakeRows = (
-  incomingStakes: StakeType[],
-  loading: boolean,
-  basePath: string,
-) => {
-  return !loading
-    ? incomingStakes?.map((incomingStake) => {
-        return {
-          cells: [
-            {
-              children: <FormattedValue format={'account'} value={incomingStake} />,
-              className: 'desktop:w-1/5',
-            },
-            {
-              children: <Currency amount={incomingStake?.amount} />,
-            },
-          ],
-        };
-      })
-    : getTableSkeletons(validatorStakeIncomingTableHead.length);
-};
-
-export const createValidatorOutgoingStakeRows = (
-  outgoingStakes: StakeType[],
-  validator: ValidatorType | undefined,
-  loading: boolean,
-  basePath: string,
-) => {
-  return !loading
-    ? outgoingStakes?.map((outgoingStake) => {
-        return {
-          cells: [
-            {
-              children: <FormattedValue format={'account'} value={outgoingStake} />,
-            },
-            {
-              children: <Currency amount={validator?.validatorWeight || 0} />,
-            },
-            {
-              children: (
-                <FormattedValue
-                  format={'percentage'}
-                  typographyProps={{ variant: 'paragraph-sm' }}
-                  value={formatCommission(validator?.commission)}
-                />
-              ),
-            },
-            {
-              children: <Currency amount={outgoingStake?.amount} />,
-            },
-          ],
-        };
-      })
-    : getTableSkeletons(validatorStakeOutgoingTableHead.length);
-};
-
-export const createValidatorBlockRows = (
-  blocks: BlockType[],
-  loading: boolean,
-  basePath: string,
-) => {
-  return !loading
-    ? blocks?.map((block) => {
-        return {
-          cells: [
-            {
-              children: (
-                <FormattedValue format={'string'} link={`/blocks/${block.id}`} value={block.id} />
-              ),
-            },
-            {
-              children: <FormattedValue format={'number'} value={block.height} />,
-            },
-            {
-              children: <FormattedValue format={'fromNow'} value={block.timestamp} />,
-            },
-            {
-              children: <FormattedValue format={'number'} value={block.numberOfTransactions} />,
-            },
-            {
-              children: <Currency amount={block.reward ?? 0} decimals={2} />,
-            },
-          ],
-        };
-      })
-    : getTableSkeletons(validatorBlocksTableHead.length);
-};
-
-export const createValidatorEventsRow = (events: EventsType[], loading: boolean) => {
-  return !loading
-    ? events?.map((event) => {
-        return {
-          rowDetails: (
-            <JsonViewer
-              className={'!border-0'}
-              copy
-              data={{ data: event.data as unknown as DataType }}
-              startOpen
-            />
-          ),
-          cells: [
-            {
-              children: (
-                <FormattedValue
-                  format={'fromNow'}
-                  typographyProps={{ color: 'onBackground' }}
-                  value={event.block.timestamp}
-                />
-              ),
-              className: 'desktop:w-1/5',
-            },
-            {
-              children: <FormattedValue format={'number'} value={event.block.height} />,
-            },
-            {
-              children: <FormattedValue format={'string'} value={event.module} />,
-            },
-            {
-              children: <FormattedValue format={'string'} value={event.name} />,
-              className: 'desktop:w-1/5',
-            },
-            {
-              children: <Currency amount={event.data.amount || 0} decimals={5} />,
-            },
-          ],
-        };
-      })
-    : getTableSkeletons(validatorBlocksTableHead.length);
-};
-
-export const createStakesOverviewRows = (
-  stakes: TransactionType[],
-  loading: boolean,
-  basePath: string,
-) => {
-  return !loading
-    ? stakes?.map((stake) => {
-        return {
-          cells: [
-            {
-              children: (
-                <KeyValueComponent
-                  contentValue={
-                    <FormattedValue
-                      format={'address'}
-                      link={`/transactions/${stake.id}`}
-                      value={shortString(stake?.id, 12, 'center')}
-                    />
-                  }
-                  keyValue={<StatusIcon status={stake.executionStatus} />}
-                />
-              ),
-            },
-            {
-              children: <FormattedValue format={'fromNow'} value={stake.block.timestamp} />,
-            },
-            {
-              children: <FormattedValue format={'account'} value={stake?.sender} />,
-            },
-            {
-              children: (
-                <>
-                  <div className="flex flex-col ">
-                    {stake?.params?.stakes?.length > 1
-                      ? stake?.params?.stakes?.map((param: any) => {
-                          const amount = param?.amount;
-                          const color = amount > 0 ? 'success' : 'error';
-                          return (
-                            <div
-                              className="flex items-center justify-between gap-8 w-72 -m-0.5"
-                              key={param?.validatorAddress}
-                            >
-                              <FormattedValue
-                                format={'account'}
-                                value={{ address: param?.validatorAddress, name: param?.name }}
-                              />
-                              <Currency
-                                amount={amount}
-                                className="text-right self-end"
-                                color={color}
-                                decimals={2}
-                                fontWeight="normal"
-                                variant="paragraph-sm"
-                              />
-                            </div>
-                          );
-                        })
-                      : stake?.params?.stakes?.map((param: any) => {
-                          const amount = param?.amount;
-                          const color = amount > 0 ? 'success' : 'error';
-                          return (
-                            <div
-                              className="flex items-center justify-between gap-8 w-72 -m-0.5"
-                              key={param?.validatorAddress}
-                            >
-                              <FormattedValue
-                                format={'account'}
-                                value={{ address: param?.validatorAddress, name: param?.name }}
-                              />
-                              <Currency
-                                amount={amount}
-                                className="text-right self-end"
-                                color={color}
-                                decimals={2}
-                                fontWeight="normal"
-                                variant="paragraph-sm"
-                              />
-                            </div>
-                          );
-                        })}
-                  </div>
-                </>
-              ),
-            },
-          ],
-        };
-      })
-    : getTableSkeletons(stakesOverviewTableHead.length);
-};
-
-export const createBlockRows = (
-  blocks: BlockType[],
-  loading: boolean,
-  copyTooltipText: string,
-  setCopyTooltipText: (text: string) => void,
-  basePath: string,
-) => {
-  const columnCount = blockTableHead(() => '', '', '').length;
-  return !loading
-    ? blocks?.map((block) => {
-        return {
-          cells: [
-            {
-              children: (
-                <KeyValueComponent
-                  contentValue={
-                    <FormattedValue
-                      copy
-                      format={'address'}
-                      link={`/blocks/${block.id}`}
-                      showCopyOnHover
-                      value={block.id}
-                    />
-                  }
-                  keyValue={<StatusIcon connected={block.isFinal} />}
-                />
-              ),
-              className: 'group/child',
-            },
-            {
-              children: (
-                <FormattedValue
-                  copy
-                  format={'number'}
-                  showCopyOnHover
-                  typographyProps={{ color: 'onBackgroundLow' }}
-                  value={block.height}
-                />
-              ),
-              className: 'group/child min-w-[120px]',
-            },
-            {
-              children: <FormattedValue format={'fromNow'} value={block.timestamp} />,
-            },
-            {
-              children: <FormattedValue format={'account'} value={block.generator} />,
-            },
-            {
-              children: (
-                <FormattedValue format={'address'} value={getSeedRevealFromAssets(block.assets)} />
-              ),
-            },
-            {
-              children: (
-                <FormattedValue format={'number'} value={block.numberOfTransactions || 0} />
-              ),
-            },
-            {
-              children: <FormattedValue format={'number'} value={block.numberOfEvents || 0} />,
-            },
-            {
-              children: <FormattedValue format={'number'} value={block.numberOfAssets || 0} />,
-            },
-          ],
-        };
-      })
-    : getTableSkeletons(columnCount);
-};
-
-export const createFavouritesRows = (
-  favourites: FavouriteType[],
-  loading: boolean,
-  basePath: string,
-  removeFavourite: (favourite: FavouriteType) => void,
-) => {
-  return !loading
-    ? favourites?.map((fav) => {
-        return {
-          cells: [
-            {
-              children: <FormattedValue format={'account'} value={fav} />,
-            },
-            {
-              children: (
-                <IconButton
-                  className={'group-hover:block desktop:hidden'}
-                  icon={'Trash'}
-                  onClick={() => removeFavourite(fav)}
-                  variant={'quaternary'}
-                />
-              ),
-              className: 'w-iconButtonWidth',
-            },
-          ],
-        };
-      })
-    : getTableSkeletons(favouritesTableHead.length);
-};
-
-export const createAccountsRows = (accounts: AccountType[], loading: boolean, basePath: string) => {
-  return !loading
-    ? accounts?.map((account, index) => {
-        return {
-          cells: [
-            {
-              //mock_data
-              children: <Typography>{index + 1}</Typography>,
-            },
-            {
-              //mock_data
-              children: (
-                <FormattedValue
-                  accountProps={{
-                    addressColor: 'onBackgroundLow',
-                    addressVariant: 'caption',
-                    nameColor: 'onBackgroundMedium',
-                    nameFontWeight: 'semibold',
-                    nameVariant: 'paragraph-sm',
-                  }}
-                  format={'account'}
-                  value={account}
-                />
-              ),
-            },
-            {
-              children: account?.description ? (
-                <Typography color={'onBackgroundMedium'} variant={'caption'}>
-                  {account?.description}
-                </Typography>
-              ) : null,
-            },
-            {
-              //mock_data
-              children: (
-                <div className="flex flex-col items-end">
-                  <Currency amount={account?.totalBalance} className="font-semibold" decimals={0} />
-                </div>
-              ),
-            },
-            {
-              //mock_data
-              children: (
-                <div className="flex flex-col items-end">
-                  <Currency
-                    amount={account?.availableBalance}
-                    className="font-semibold"
-                    decimals={0}
-                  />
-                  <FormattedValue
-                    format={'percentage'}
-                    value={Number(
-                      (
-                        (Number(account?.availableBalance || 0) /
-                          Number(account?.totalBalance || 1)) *
-                        100
-                      ).toFixed(2),
-                    )}
-                  />
-                </div>
-              ),
-            },
-            {
-              //mock_data
-              children: (
-                <div className="flex flex-col items-end">
-                  <Currency
-                    amount={account?.lockedBalance}
-                    className="font-semibold"
-                    decimals={0}
-                  />
-                  <FormattedValue
-                    format={'percentage'}
-                    value={Number(
-                      (
-                        (Number(account?.lockedBalance || 0) / Number(account?.totalBalance || 1)) *
-                        100
-                      ).toFixed(2),
-                    )}
-                  />
-                </div>
-              ),
-            },
-            {
-              //mock_data
-              children: <FormattedValue format={'percentage'} value={0} />,
-              className: 'text-right',
-            },
-          ],
-        };
-      })
-    : getTableSkeletons(6);
-};
-
-export const createChainRows = (chains: CombinedAppsType[], loading: boolean, basePath: string) => {
-  return !loading
-    ? chains?.map((chain) => {
-        return {
-          cells: [
-            {
-              children: (
-                <ImageName imageUrl={chain.logo?.png} name={chain.displayName ?? chain.chainName} />
-              ),
-              className: 'w-72',
-            },
-            {
-              children: <StatusBadge status={chain.status} />,
-              className: 'w-44',
-            },
-            {
-              children: <Currency amount={chain.escrowedKLY} decimals={3} />,
-              className: 'text-right w-52',
-            },
-            chain.logo
-              ? {
-                  children: (
-                    <FlexGrid mobileDirection={'row'} gap={'lg'}>
-                      <Link
-                        className={'contents'}
-                        basePath={basePath}
-                        href={`/chains/${chain.chainID}`}
-                      >
-                        <Button align={'right'} label={'Details'} variant={'bordered'} />
-                      </Link>
-                      <IconButton align={'none'} icon={'LinkExternal'} variant={'quaternary'} />
-                    </FlexGrid>
-                  ),
-                }
-              : { children: null },
-          ],
-        };
-      })
-    : getTableSkeletons(chainsTableHead.length);
-};
-
-export const createUserDetailsTokensRow = (
-  token: TokenType[],
-  chain: ChainType,
-  loading: boolean,
-) => {
-  return !loading
-    ? token?.map((token) => {
-        const totalBalance =
-          Number(token.availableBalance) + Number(token.lockedBalances?.[0]?.amount ?? 0);
-        const availablePercentage =
-          totalBalance > 0
-            ? ((Number(token.availableBalance) / totalBalance) * 100).toFixed(2)
-            : '0.00';
-        const lockedPercentage =
-          totalBalance > 0
-            ? ((Number(token.lockedBalances?.[0]?.amount ?? 0) / totalBalance) * 100).toFixed(2)
-            : '0.00';
-
-        return {
-          cells: [
-            {
-              children: (
-                <TokenCard
-                  chainImage={
-                    'https://images.crunchbase.com/image/upload/c_pad,h_170,w_170,f_auto,b_white,q_auto:eco,dpr_1/iajdm4uwsshvi1d4dt7g'
-                  }
-                  image={'https://cdn.pixabay.com/photo/2023/10/17/17/01/cat-8321993_1280.jpg'}
-                  name={'Monkeyz'}
-                  symbol={'MON'}
-                />
-              ),
-            },
-            {
-              children: (
-                <div className="flex flex-col">
-                  <Currency amount={totalBalance} decimals={0} fontWeight={'semibold'} />
-                  <Currency
-                    amount={Number(token.availableBalance) * 2}
-                    className="text-onBackgroundLow text-caption"
-                    decimals={2}
-                    sign={'$'}
-                  />
-                </div>
-              ),
-            },
-            {
-              children: (
-                <div className="flex flex-col ">
-                  <Currency amount={token.availableBalance} decimals={0} fontWeight={'semibold'} />
-                  <FormattedValue format={'percentage'} value={availablePercentage} />
-                </div>
-              ),
-            },
-            {
-              children: (
-                <div className="flex flex-col ">
-                  <Currency
-                    amount={token.lockedBalances?.[0]?.amount ?? 0}
-                    decimals={0}
-                    fontWeight={'semibold'}
-                  />
-                  <FormattedValue format={'percentage'} value={lockedPercentage} />
-                </div>
-              ),
-            },
-            {
-              children: <FormattedValue format={'number'} value={2} />,
-            },
-            {
-              children: (
-                <ImageName
-                  imageUrl={chain.logo.png ?? ''}
-                  name={chain.displayName ?? chain.chainName ?? ''}
-                />
-              ),
-            },
-          ],
-        };
-      })
-    : getTableSkeletons(6);
-};
-
-export const createTokensRows = (tokens: TokenType[], loading: boolean, basePath: string) => {
-  return !loading
-    ? tokens?.map((token) => {
-        return {
-          cells: [
-            {
-              children: (
-                <Link basePath={basePath} href={`/tokens/${token.tokenId}`}>
-                  <TokenCard
-                    chainImage={
-                      'https://images.crunchbase.com/image/upload/c_pad,h_170,w_170,f_auto,b_white,q_auto:eco,dpr_1/iajdm4uwsshvi1d4dt7g'
-                    }
-                    image={'https://cdn.pixabay.com/photo/2023/10/17/17/01/cat-8321993_1280.jpg'}
-                    name={'Monkeyz'}
-                    symbol={'MON'}
-                  />
-                </Link>
-              ),
-            },
-            {
-              children: (
-                <ImageName
-                  imageUrl={
-                    'https://images.crunchbase.com/image/upload/c_pad,h_170,w_170,f_auto,b_white,q_auto:eco,dpr_1/iajdm4uwsshvi1d4dt7g'
-                  }
-                  name={'Klayr-main'}
-                />
-              ),
-            },
-            {
-              children: (
-                <KeyValueComponent
-                  contentValue={
-                    <FormattedValue
-                      format={'string'}
-                      typographyProps={{ color: 'onBackgroundHigh' }}
-                      value={'Testnet'}
-                    />
-                  }
-                  keyValue={<StatusIcon status={'successful'} />}
-                />
-              ),
-            },
-            {
-              children: (
-                <FormattedValue
-                  format={'string'}
-                  value={'The description of the token and its purpose'}
-                />
-              ),
-            },
-          ],
-        };
-      })
-    : getTableSkeletons(4);
-};
-
-export const createNodesRows = (nodes: NodeType[], loading: boolean) => {
-  return !loading
-    ? nodes?.map((node) => {
-        return {
-          cells: [
-            {
-              children: <FormattedValue format={'string'} value={node.ipAddress} />,
-            },
-            {
-              children: <FormattedValue format={'string'} value={node.port} />,
-            },
-            {
-              children: (
-                <div className="flex gap-2 items-center">
-                  <div className="flex h-6 w-6 bg-white rounded-full items-center justify-center">
-                    <div className="flex h-3 w-3 bg-error rounded-full"></div>
-                  </div>
-                  <Typography
-                    color={'onBackgroundMedium'}
-                    fontWeight={'semibold'}
-                    variant={'paragraph-sm'}
-                  >
-                    {'Japan'}
-                  </Typography>
-                </div>
-              ),
-            },
-            {
-              children: <FormattedValue format={'number'} value={node.options.blockVersion} />,
-            },
-            {
-              children: <FormattedValue format={'number'} value={node.options.height} />,
-            },
-            {
-              children: <StatusBadge status={'online'} />,
-            },
-          ],
-        };
-      })
-    : getTableSkeletons(6);
-};
-
-export const createNftsRows = (nfts: NftType[], loading: boolean) => {
-  return !loading
-    ? nfts?.map((nft) => {
-        return {
-          cells: [
-            {
-              children: <ImageName imageUrl={nft.image ?? ''} name={nft.name ?? ''} />,
-            },
-            {
-              children: <FormattedValue format={'string'} value={nft.collection} />,
-            },
-            {
-              children: (
-                <div className="flex flex-col">
-                  <Currency amount={nft.price ?? 0} decimals={3} />
-                  <Currency
-                    amount={Number(nft.price) * 0.7}
-                    color="onBackgroundLow"
-                    decimals={2}
-                    sign={'$'}
-                  />
-                </div>
-              ),
-            },
-            {
-              children: <StatusBadge status={nft.status || ''} />,
-            },
-            {
-              children: <FormattedValue format={'string'} value={`#${nft.rarityRank}`} />,
-            },
-            {
-              children: <ImageName imageUrl={nft.chainImage ?? ''} name={nft.chain ?? ''} />,
-            },
-          ],
-        };
-      })
-    : getTableSkeletons(6);
-};
-
-export const createNftsPageRows = (nfts: NftType[], loading: boolean) => {
-  return !loading
-    ? nfts?.map((nft) => {
-        return {
-          cells: [
-            {
-              children: <ImageName imageUrl={nft.image ?? ''} name={nft.name ?? ''} />,
-            },
-            {
-              children: <FormattedValue format={'string'} value={nft.collection} />,
-            },
-            {
-              children: <StatusBadge status={nft.status || ''} />,
-            },
-            {
-              children: (
-                <div className="flex flex-col">
-                  {/* <Currency amount={nft.price ?? 0} decimals={3} symbol={'KLY'} /> */}
-                  <Currency
-                    amount={Number(nft.price) * 0.7}
-                    color="onBackgroundLow"
-                    decimals={2}
-                    sign={'$'}
-                  />
-                </div>
-              ),
-            },
-            {
-              children: <ImageName imageUrl={nft.chainImage ?? ''} name={nft.chain ?? ''} />,
-            },
-          ],
-        };
-      })
-    : getTableSkeletons(5);
-};
